#876 - Upgrade to Java 21 & change key to Long

+++ jpa/boot-read-replica-postgresql/src/main/java/com/example/demo/readreplica/controller/ArticleController.java
@@ -29,7 +29,7 @@ public class ArticleController {
 
     @PostMapping("/")
     public ResponseEntity<Object> saveArticle(@RequestBody ArticleDTO articleDTO) {
-        Integer articleId = this.articleService.saveArticle(articleDTO);
+        Long articleId = this.articleService.saveArticle(articleDTO);
         return ResponseEntity.created(URI.create("/articles/" + articleId)).build();
     }
 }
