+++ jpa/boot-read-replica-postgresql/src/main/java/com/example/demo/readreplica/repository/CommentRepository.java
@@ -3,4 +3,4 @@ package com.example.demo.readreplica.repository;
 import com.example.demo.readreplica.entities.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommentRepository extends JpaRepository<Comment, Integer> {}
+public interface CommentRepository extends JpaRepository<Comment, Long> {}
