+++ jpa/boot-read-replica-postgresql/src/main/java/com/example/demo/readreplica/service/ArticleService.java
@@ -22,7 +22,7 @@ public class ArticleService {
     }
 
     @Transactional
-    public Integer saveArticle(ArticleDTO articleDTO) {
+    public Long saveArticle(ArticleDTO articleDTO) {
         Article article = convertToArticle(articleDTO);
         Article savedArticle = this.articleRepository.save(article);
         return savedArticle.getId();
